refactor: entry-sectionをcomponents直下へ移設しsections/を削除
