+++ src/components/entry-section.tsx
