+++ src/features/income/index.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { SectionShell, EntryRow } from '@/components/sections/entry-section'
+import { SectionShell, EntryRow } from '@/components/entry-section'
 import { DeleteButton } from '@/components/ui/delete-button'
 import { AddEntryModal } from '@/features/add-entry'
 import { EditModal } from '@/features/edit-entry'
